requirements.txt
redis>=4.5.0
numpy>=1.22.0
asyncio>=3.4.3
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-timeout>=2.1.0
pyyaml>=6.0
jsonschema>=4.17.0
